ft_bzero

+++ libft/ft_memset.c
@@ -1,8 +1,8 @@
 #include "libft.h"
 
-void	*ft_memset(void *s, int c, int n)
+void	*ft_memset(void *s, int c, size_t n)
 {
-	int	i;
+	size_t	i;
 
 	i = 0;
 	if (!s)
